transferred functionality according to client and trader
